refactor: handle detox in release & debug
