+++ app.config.ts
@@ -12,26 +12,39 @@ let Config = {
   androidPackageName: 'com.tsyirvo.rnstarter',
 };
 
-if (appEnv === 'development') {
-  Config.appName = 'Dev';
-  Config.iosBundleId = 'com.tsyirvo.rnstarter.development';
-  Config.androidPackageName = 'com.tsyirvo.rnstarter.development';
-} else if (appEnv === 'test') {
-  Config.appName = 'Test';
-  Config.iosBundleId = 'com.tsyirvo.rnstarter.test';
-  Config.androidPackageName = 'com.tsyirvo.rnstarter.test';
-} else if (appEnv === 'staging') {
-  Config.appName = 'Staging';
-  Config.iosBundleId = 'com.tsyirvo.rnstarter.staging';
-  Config.androidPackageName = 'com.tsyirvo.rnstarter.staging';
+switch (appEnv) {
+  case 'development':
+    Config.appName = 'Dev';
+    Config.iosBundleId = 'com.tsyirvo.rnstarter.development';
+    Config.androidPackageName = 'com.tsyirvo.rnstarter.development';
+    break;
+  case 'test:debug':
+    Config.appName = 'TestDebug';
+    Config.iosBundleId = 'com.tsyirvo.rnstarter.test.debug';
+    Config.androidPackageName = 'com.tsyirvo.rnstarter.test.debug';
+    break;
+  case 'test:release':
+    Config.appName = 'TestRelease';
+    Config.iosBundleId = 'com.tsyirvo.rnstarter.test.release';
+    Config.androidPackageName = 'com.tsyirvo.rnstarter.test.release';
+    break;
+  case 'staging':
+    Config.appName = 'Staging';
+    Config.iosBundleId = 'com.tsyirvo.rnstarter.staging';
+    Config.androidPackageName = 'com.tsyirvo.rnstarter.staging';
+    break;
+  default:
+    break;
 }
+
+const isTestEnv = appEnv === 'test:debug' || appEnv === 'test:release';
 
 export default ({ config }: ConfigContext): ExpoConfig => ({
   ...config,
   name: Config.appName,
   slug: 'rn-starter',
   plugins:
-    appEnv === 'test' && config.plugins
+    isTestEnv && config.plugins
       ? [...config.plugins, '@config-plugins/detox']
       : config.plugins,
   ios: {
