--- e2e/builds/Test.app/EXConstants.bundle/app.config
@@ -1,1 +1,0 @@
-{"name":"Test","slug":"rn-starter-expo","version":"1.0.0","runtimeVersion":"1.0.0","jsEngine":"hermes","orientation":"portrait","icon":"./assets/icon.png","userInterfaceStyle":"light","splash":{"image":"./assets/splash.png","resizeMode":"contain","backgroundColor":"#ffffff"},"updates":{"fallbackToCacheTimeout":0},"assetBundlePatterns":["**/*"],"plugins":["sentry-expo","expo-community-flipper","@config-plugins/detox"],"ios":{"supportsTablet":false,"buildNumber":"1","bundleIdentifier":"com.tsyirvo.rnstarter.test"},"android":{"versionCode":1,"adaptiveIcon":{"foregroundImage":"./assets/adaptive-icon.png","backgroundColor":"#FFFFFF"},"package":"com.tsyirvo.rnstarter.test"},"extra":{"eas":{"projectId":"b8ae205e-a96b-4b22-a147-768e74f54b37","sentryDsn":"https://ee3c8e32d176431e9dea6e297d29c0d2@o1097704.ingest.sentry.io/6119381","mixpanelToken":"7ef111931c8812b0d0ff45887306da2f","firebaseConfig":{"apiKey":"AIzaSyAJ4m7ncD44dycWVywZ4mITkDlKl2hWJFk","authDomain":"rn-starter-expo.firebaseapp.com","projectId":"rn-starter-expo","storageBucket":"rn-starter-expo.appspot.com","messagingSenderId":"972369654016","appId":"1:972369654016:web:db7a00dddf02a9c00b43e5","measurementId":"G-YEE0VJ3T8K"},"appEnv":"test","appName":"Test","iosBundleId":"com.tsyirvo.rnstarter.test","androidPackageName":"com.tsyirvo.rnstarter.test"}},"sdkVersion":"46.0.0","platforms":["ios","android","web"],"currentFullName":"@tsyirvo/rn-starter-expo","originalFullName":"@tsyirvo/rn-starter-expo"}
