--- e2e/builds/Test.app/EXUpdates.bundle/app.manifest
@@ -1,1 +1,0 @@
-{"id":"e35d5cfb-b851-4e7f-88e5-0ce823612daf","commitTime":1665935970790,"assets":[{"name":"WorkSans-Light","type":"ttf","scale":1,"packagerHash":"f6c8351e63aa1c05032ed3e2b78cd980","subdirectory":"/assets/src/assets/fonts","nsBundleDir":"assets/src/assets/fonts","nsBundleFilename":"WorkSans-Light"},{"name":"WorkSans-Regular","type":"ttf","scale":1,"packagerHash":"a2fa05d4ae83871481e0038529a017da","subdirectory":"/assets/src/assets/fonts","nsBundleDir":"assets/src/assets/fonts","nsBundleFilename":"WorkSans-Regular"},{"name":"WorkSans-Medium","type":"ttf","scale":1,"packagerHash":"80eff1de1e9871230ecd0ebbbb4efd7f","subdirectory":"/assets/src/assets/fonts","nsBundleDir":"assets/src/assets/fonts","nsBundleFilename":"WorkSans-Medium"},{"name":"WorkSans-Bold","type":"ttf","scale":1,"packagerHash":"17fd2e111396bdc778130cbc5487a1bb","subdirectory":"/assets/src/assets/fonts","nsBundleDir":"assets/src/assets/fonts","nsBundleFilename":"WorkSans-Bold"},{"name":"back-icon","type":"png","scale":1,"packagerHash":"7d40544b395c5949f4646f5e150fe020","subdirectory":"/assets/node_modules/@react-navigation/elements/src/assets","nsBundleDir":"assets/node_modules/@react-navigation/elements/src/assets","nsBundleFilename":"back-icon"},{"name":"back-icon","type":"png","scale":2,"packagerHash":"cdd04e13d4ec83ff0cd13ec8dabdc341","subdirectory":"/assets/node_modules/@react-navigation/elements/src/assets","nsBundleDir":"assets/node_modules/@react-navigation/elements/src/assets","nsBundleFilename":"back-icon@2x"},{"name":"back-icon","type":"png","scale":3,"packagerHash":"a132ecc4ba5c1517ff83c0fb321bc7fc","subdirectory":"/assets/node_modules/@react-navigation/elements/src/assets","nsBundleDir":"assets/node_modules/@react-navigation/elements/src/assets","nsBundleFilename":"back-icon@3x"},{"name":"back-icon-mask","type":"png","scale":1,"packagerHash":"5223c8d9b0d08b82a5670fb5f71faf78","subdirectory":"/assets/node_modules/@react-navigation/elements/src/assets","nsBundleDir":"assets/node_modules/@react-navigation/elements/src/assets","nsBundleFilename":"back-icon-mask"},{"name":"header","type":"jpeg","scale":1,"packagerHash":"08cb578eb67a521bd6e08e7190159678","subdirectory":"/assets/src/assets/images","nsBundleDir":"assets/src/assets/images","nsBundleFilename":"header"},{"name":"header","type":"jpeg","scale":2,"packagerHash":"d59bce655e31abdd21ffecf100931185","subdirectory":"/assets/src/assets/images","nsBundleDir":"assets/src/assets/images","nsBundleFilename":"header@2x"},{"name":"header","type":"jpeg","scale":3,"packagerHash":"2bf12faf9bfe04a2df940f846513ce4c","subdirectory":"/assets/src/assets/images","nsBundleDir":"assets/src/assets/images","nsBundleFilename":"header@3x"}]}
