--- e2e/builds/Test.app/PkgInfo
@@ -1,1 +1,0 @@
-APPL????
