+++ e2e/utils/helpers.ts
@@ -1,0 +1,14 @@
+import { packagerUrl, START_DELAY } from './constants';
+
+export const sleep = async (ms: number) =>
+  new Promise((resolve) => setTimeout(resolve, ms));
+
+export const startApp = async (config?: Detox.DeviceLaunchAppConfig) => {
+  await device.launchApp({
+    newInstance: true,
+    url: packagerUrl,
+    ...config,
+  });
+
+  await sleep(START_DELAY);
+};
