+++ package.json
@@ -19,10 +19,11 @@
     "Development Build"
   ],
   "license": "ISC",
-  "type": "module",
   "scripts": {
     "start:dev": "APP_ENV=development expo start --dev-client",
     "start:test": "APP_ENV=test expo start --dev-client --no-dev --minify",
+    "start:test:android": "APP_ENV=test expo start --android",
+    "start:test:ios": "APP_ENV=test expo start --ios",
     "start:staging": "APP_ENV=staging expo start --no-dev --minify",
     "start:production": "APP_ENV=production expo start --no-dev --minify",
     "//// BUILDS ////": "",
@@ -36,8 +37,12 @@
     "test:coverage": "jest",
     "e2e:build:ios:debug": "detox build --configuration ios:debug",
     "e2e:test:ios:debug": "detox test --configuration ios:debug",
-    "e2e:build:android": "detox build --configuration android",
-    "e2e:test:android": "detox test --configuration android",
+    "e2e:build:ios:release": "detox build --configuration ios:release",
+    "e2e:test:ios:release": "detox test --configuration ios:release",
+    "e2e:build:android:debug": "detox build --configuration android:debug",
+    "e2e:test:android:debug": "detox test --configuration android:debug",
+    "e2e:build:android:release": "detox build --configuration android:release",
+    "e2e:test:android:release": "detox test --configuration android:release",
     "//// QUALITY ////": "",
     "tsc": "tsc",
     "lint": "eslint ./ --ext .js,.ts,.tsx --fix --ignore-path .gitignore",
@@ -66,7 +71,7 @@
     "expo-application": "~4.2.2",
     "expo-asset": "~8.6.1",
     "expo-constants": "~13.2.4",
-    "expo-dev-client": "~1.2.1",
+    "expo-dev-client": "~1.3.1",
     "expo-dev-menu": "1.2.1",
     "expo-device": "~4.3.0",
     "expo-font": "~10.2.0",
