+++ scripts/detox/move-android-release-apk.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+mv "$PWD/e2e/builds/android-release/release/app-release.apk" "$PWD/e2e/builds/"
+mv "$PWD/e2e/builds/android-release/androidTest/release/app-release-androidTest.apk" "$PWD/e2e/builds/"
